requirements.txt

Flask
darkskylib
itsdangerous
python-yr
RPi.GPIO
Adafruit-DHT
uWSGI
pika
